+++ NEWS.md
@@ -1,5 +1,8 @@
 # Changes in mirt 1.30
 
+- `weights` argument added to `estfun.AllModelClass` to allow for the inclusion
+  of `survey.weights` to calculate the scores
+
 - `DIF()` now simplifies the output by default rather than returning lists from `anova()`. Wald tests are always simplified
 
 - Where applicable, RMSEA statistics are reported in `itemfit()` for tests that return suitable